Add a default sound/GIF and support multiple Financial Force tabs at once
The first part of this commit adds a simple GIF and sound to be used when the user has not specified their own. Most of the commit implements a throttle on state changes for a content script context so a change event is only sent if the state is dirtied. This allows the tab with the most recent state change to win out.